Refactor engine opening book
